requirements.txt
flask
scikit-learn
pandas
numpy
pytesseract
pillow
gunicorn
python-dotenv
requests